+++ webview-ui/src/components/mcp/configuration/tabs/installed/server-row/McpResourceRow.tsx
@@ -1,4 +1,4 @@
-import { McpResource, McpResourceTemplate } from "../../../../src/shared/mcp"
+import { McpResource, McpResourceTemplate } from "../../../../../../../../src/shared/mcp"
 
 type McpResourceRowProps = {
 	item: McpResource | McpResourceTemplate
